+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-28T16:26:12.171Z",
+    "2026-04-29T12:22:32.001Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {